Check for deploy.json before removing, set proper file perms

+++ deploy-build.sh
@@ -4,7 +4,9 @@
 
 # Remove any existing deploy trigger to prevent concurrent operations in the
 # case where multiple projects are being deployed at once.
-rm /var/repos/wsuwp-deployment/deploy.json
+if [ -f "/var/repos/wsuwp-deployment/deploy.json" ]; then
+  rm /var/repos/wsuwp-deployment/deploy.json
+fi
 
 # If a deployment is not yet configured, expect a third argument containing
 # the repository URL. We'll need to configure private repositories manually.
@@ -44,5 +46,6 @@ grunt
 
 # Tell cron that we're again ready for deploy.
 touch /var/repos/wsuwp-deployment/deploy.json
+chmod 664 /var/repos/wsuwp-deployment/deploy.json
 
 exit 1
